+++ pkg/storage/chunk/schema.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/go-kit/log/level"
 	jsoniter "github.com/json-iterator/go"
@@ -968,9 +967,6 @@ func (v11Entries) GetLabelNamesForSeries(bucket Bucket, seriesID []byte) ([]Inde
 	}, nil
 }
 
-// v12Entries adds ChunkPathShardFactor and ChunkPathPeriod as SchemaConfig values
 type v12Entries struct {
 	v11Entries
-	ChunkPathShardFactor uint64
-	ChunkPathPeriod      time.Duration
 }