+++ pkg/storage/chunk/schema_config.go
@@ -36,15 +36,13 @@ var (
 
 // PeriodConfig defines the schema and tables to use for a period of time
 type PeriodConfig struct {
-	From                 DayTime             `yaml:"from"`         // used when working with config
-	IndexType            string              `yaml:"store"`        // type of index client to use.
-	ObjectType           string              `yaml:"object_store"` // type of object client to use; if omitted, defaults to store.
-	Schema               string              `yaml:"schema"`
-	IndexTables          PeriodicTableConfig `yaml:"index"`
-	ChunkTables          PeriodicTableConfig `yaml:"chunks"`
-	RowShards            uint32              `yaml:"row_shards"`
-	ChunkPathShardFactor uint64              `yaml:"chunk_path_shard_factor"`
-	ChunkPathPeriod      time.Duration       `yaml:"chunk_path_period"`
+	From        DayTime             `yaml:"from"`         // used when working with config
+	IndexType   string              `yaml:"store"`        // type of index client to use.
+	ObjectType  string              `yaml:"object_store"` // type of object client to use; if omitted, defaults to store.
+	Schema      string              `yaml:"schema"`
+	IndexTables PeriodicTableConfig `yaml:"index"`
+	ChunkTables PeriodicTableConfig `yaml:"chunks"`
+	RowShards   uint32              `yaml:"row_shards"`
 }
 
 // DayTime is a model.Time what holds day-aligned values, and marshals to/from
@@ -132,26 +130,6 @@ func defaultRowShards(schema string) uint32 {
 	}
 }
 
-// ChunkPathPeriod is introduced as a SchemaConfig value in Schema "v12"
-func defaultChunkPathPeriod(schema string) time.Duration {
-	switch schema {
-	case v12:
-		return 1 * time.Hour
-	default:
-		return 0
-	}
-}
-
-// ChunkPathShardFactor is introduced as a SchemaConfig value in Schema "v12"
-func defaultChunkPathShardFactor(schema string) uint64 {
-	switch schema {
-	case v12:
-		return 2
-	default:
-		return 0
-	}
-}
-
 // ForEachAfter will call f() on every entry after t, splitting
 // entries if necessary so there is an entry starting at t
 func (cfg *SchemaConfig) ForEachAfter(t model.Time, f func(config *PeriodConfig)) {
@@ -223,10 +201,7 @@ func (cfg PeriodConfig) CreateSchema() (BaseSchema, error) {
 		} else if cfg.Schema == "v11" {
 			return newSeriesStoreSchema(buckets, v11Entries{v10}), nil
 		} else { // v12
-			if cfg.ChunkPathPeriod == 0 || cfg.ChunkPathShardFactor == 0 {
-				return nil, fmt.Errorf("must configure chunk_path_shard_factor and chunk_path_period to values > 0 for schema (%s)", cfg.Schema)
-			}
-			return newSeriesStoreSchema(buckets, v12Entries{v11Entries{v10}, cfg.ChunkPathShardFactor, cfg.ChunkPathPeriod}), nil
+			return newSeriesStoreSchema(buckets, v12Entries{v11Entries{v10}}), nil
 		}
 	default:
 		return nil, errInvalidSchemaVersion
@@ -245,12 +220,6 @@ func (cfg PeriodConfig) createBucketsFunc() (schemaBucketsFunc, time.Duration) {
 func (cfg *PeriodConfig) applyDefaults() {
 	if cfg.RowShards == 0 {
 		cfg.RowShards = defaultRowShards(cfg.Schema)
-	}
-	if cfg.ChunkPathPeriod == 0 {
-		cfg.ChunkPathPeriod = defaultChunkPathPeriod(cfg.Schema)
-	}
-	if cfg.ChunkPathShardFactor == 0 {
-		cfg.ChunkPathShardFactor = defaultChunkPathShardFactor(cfg.Schema)
 	}
 }
 
@@ -503,7 +472,7 @@ func (cfg SchemaConfig) ExternalKey(chunk Chunk) string {
 	p, err := cfg.SchemaForTime(chunk.From)
 	v, _ := p.VersionAsInt()
 	if err == nil && v >= 12 {
-		return cfg.newerExternalKey(chunk, p.ChunkPathShardFactor, p.ChunkPathPeriod)
+		return cfg.newerExternalKey(chunk)
 	} else if chunk.ChecksumSet {
 		return cfg.newExternalKey(chunk)
 	} else {
@@ -525,14 +494,6 @@ func (cfg SchemaConfig) newExternalKey(chunk Chunk) string {
 }
 
 // v12+
-func (cfg SchemaConfig) newerExternalKey(chunk Chunk, shardFactor uint64, period time.Duration) string {
-	shard := uint64(chunk.Fingerprint) % shardFactor
-	// Reduce the fingerprint into the <period> space to act as jitter
-	jitter := uint64(chunk.Fingerprint) % uint64(period)
-	// The fingerprint (hash, uniformly distributed) allows us to gradually
-	// write into the next <period>, "warming" it up in a linear fashion wrt time.
-	// This means that if we're 10% into the current period, 10% of fingerprints will
-	// be written into the next period instead.
-	prefix := (uint64(chunk.From.UnixNano()) + jitter) % uint64(period)
-	return fmt.Sprintf("%s/%x/%x/%x/%x:%x:%x", chunk.UserID, prefix, shard, uint64(chunk.Fingerprint), int64(chunk.From), int64(chunk.Through), chunk.Checksum)
+func (cfg SchemaConfig) newerExternalKey(chunk Chunk) string {
+	return fmt.Sprintf("%s/%x/%x:%x:%x", chunk.UserID, uint64(chunk.Fingerprint), int64(chunk.From), int64(chunk.Through), chunk.Checksum)
 }