+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 * [4904](https://github.com/grafana/loki/pull/4904) **bboreham**: Fixes rare race condition that could crash an ingester.
 * [4942](https://github.com/grafana/loki/pull/4942) **cyriltovena**: Allow to disable HTTP/2 for GCS.
 * [4876](https://github.com/grafana/loki/pull/4876) **trevorwhitney**: Docs: add simple, scalable example using docker-compose
-* [4857](https://github.com/grafana/loki/pull/4857) **jordanrushing**: New schema v12 including additional configuration for chunk object prefixes
+* [4857](https://github.com/grafana/loki/pull/4857) **jordanrushing**: New schema v12 changes chunk key structure
 
 # 2.4.1 (2021/11/07)